Develop circulation computation routine and change dist_vec to 1D in lumpvor2d
Developed the algorithm to compute the circulation of the airfoil using
discrete vortex method.
Additionally the dist_vec is reduced to a 1D array to reduce inefficient coding.

What may be additionally done to speed up the code, is to compute the RHS
array directly using matrix / vector multiplication. However, this was too
ambitious for my skills (apparently).