Develop run and compute_parameters function
Also added compute_panel_length method to compute the chord length.

Function is working albeit with bugs.

discreteVortexMethod.py is edited to compensate for alternative
design of the airfoil data structure.

Added another function in postProcess.py to plot the change in pressure
over x/c.

Results show that there is an error in the computation.